rename app/schema/ to app/models/
