+++ app/models/artistModel.js
